third_party/WebKit/LayoutTests/intersection-observer/zero-area-element-visible.html

<!DOCTYPE html>
<script src="../resources/intersection-observer-helper-functions.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<title>Ensure that a visible zero-area element is given the correct intersection ratio</title>

<div id='target' style='width: 0px; height: 0px'</div>"

<script>
'use strict';

async_test(t => {
  let target = document.getElementById('target');
  let entries = [];
  new IntersectionObserver(changes => {
    entries.push(...changes);
  }).observe(target);
  waitForNotification(t.step_func_done(() => {
    assert_equals(entries.length, 1);
    assert_equals(entries[0].intersectionRatio, 1);
  }));
});
</script>
